Add Keycloak user registering
